Deterministicly shuffle block index
Make it more difficult for a Downstairs attacker to tamper with a volume
by shuffling the block index Upstairs using a Pseudorandom_permutation[1].

The summary of [1] is that we want a function that is fast to compute,
is bijective over the range of 0..num_blocks, and does not require a lot
of RAM when running.

This commit uses boomphf which requires a whole lot of RAM for the
computation but runs with a much smaller RSS than the (commented out)
naive array shuffle.

It also introduces a criterion benchmark for the ShuffleContext::index
function. It's important that this function be much faster than SSD seek
time (estimated at 0.08 ms -> 0.16 ms according to Wikipedia [2]) to
make it worth it to perform this mitigation.

[1]: https://en.wikipedia.org/wiki/Pseudorandom_permutation
[2]: https://en.wikipedia.org/wiki/Hard_disk_drive_performance_characteristics